double pendulum as a systen
